bfcli: the scrollback was eating the backups
  ---[ 0x01: ten bugs, one of which had been watching us

A command is finished when the prompt is showing and a line has arrived
since we sent it. "A line has arrived" was spelled lines_.size() >
snapshot, which counts arrivals beautifully right up until the 3000-line
cap starts eating the front of the buffer. Past that the count stops
rising, then falls, and the comparison can never come true again: every
command sits there for its full 15s timeout and every backup returns
"the flight controller did not return a complete response".

The FC was answering perfectly the whole time. We were the ones who had
stopped listening. Terminal now keeps linesEver(), a counter that
neither trimming nor Ctrl+L can rewind, and the session counts that.

The test was written first and watched to fail -- fill the buffer to
setMaxLines(), send `status`, wait. See "commands with a full
scrollback" in the selftest.

  ---[ 0x02: thirty frames per second of device enumeration

Pull the USB plug and tick() called refreshPorts() on every frame, for
as long as the terminal screen stayed up: opendir /dev, stat every node,
walk six levels of sysfs per port, resolve /dev/serial/by-id, then mark
the screen dirty so the panel earned a full blit for its trouble. That
is a remarkable amount of electricity to spend rediscovering the same
three device nodes on a CM0. The unplug is handled once now.

  ---[ 0x03: mistaking patience for death

"15 seconds since the command was sent" and "15 seconds of silence" are
different claims, and only the second one means the FC is gone. A `dump
all` down the Grove UART at 9600 baud is half a minute of perfectly
healthy bytes, and we shot it and called the backup a failure. The clock
runs on the gap between bytes now, with a min() against the send so a
command typed after a long idle spell does not start life expired.

  ---[ 0x04: the smaller livestock

  * a line typed while the FC was busy was committed to history and then
    dropped on the floor with a warning. It stays on the input line
    until the wire is actually free.

  * a quick command from the menu did setText()/commit() over whatever
    you had half-typed. LineEditor::pushHistory() now feeds the history
    without going through the input line.

  * scroll_ is measured in wrapped display rows. The trim subtracted
    lines from it, so a scrolled-back viewport jumped by however much
    wrapping happened to be going on.

  * Backlight::setPercent computed percent*max/100, which on a panel
    reporting max_brightness 15 is a confident, read-back-verified write
    of zero. The comment two lines above says "never blank the panel
    entirely", which was true in the same sense as most comments.

  * present() trusted yres, line_length and the pan offset to agree with
    smem_len before memcpy'ing into the mapping. They usually do.

  * /dev/serial0, /dev/ttyAMA0 and /dev/ttyS0 are one UART wearing three
    hats; dedup compared the names instead of where they point, so the
    picker offered the same header three times.

  * the completion popup computed its height for five columns and then
    drew four, quietly losing the tail of the list.

  * a capture inherited errorCount from whatever restore ran before it.

  ---[ 0x05: paperwork

175 checks, 0 failures, host and aarch64, warning-free at -Wall -Wextra.
Same suite plus a live sim run under -fsanitize=address,undefined: also
clean. Three checks are new and one of them exists specifically because
it failed first.

+++ src/bfsession.h
@@ -94,7 +94,7 @@ private:
 
     uint64_t lastByteMs_ = 0;
     uint64_t commandSentMs_ = 0;
-    size_t linesAtSend_ = 0;
+    uint64_t linesAtSend_ = 0;
     uint64_t connectStartMs_ = 0;
     int cliAttempts_ = 0;
     bool linkLost_ = false;
